backend/pyproject.toml
[project]
name = "wiki-workspace-backend"
version = "0.1.0"
description = "Wiki Workspace backend — FastAPI + Pydantic v2 AI proxy (Phase 2 PoC)."
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "pydantic>=2.7",
    "python-dotenv>=1.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "pyright>=1.1",
    "ruff>=0.4",
]
gemini = [
    "google-generativeai>=0.7",
]
vllm = [
    "openai>=1.0",
    "transformers>=4.40",
    "jinja2>=3.1",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"
# Vendored upstream modules are DO-NOT-EDIT (see
# app/services/providers/_vendor/__init__.py). Skip lint on them.
extend-exclude = ["app/services/providers/_vendor"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.pyright]
include = ["app"]
pythonVersion = "3.11"
typeCheckingMode = "standard"
reportMissingImports = "warning"
